Refactor A2A REST routing modules
