+++ src/gateway/rest/mod.rs
@@ -1,0 +1,4 @@
+// Copyright (C) 2026 Impala Systems, Inc.
+// SPDX-License-Identifier: AGPL-3.0-only
+
+pub mod a2a;
